Improved messages Prevents OCR from happening on HEIC images

+++ ocr.js
@@ -1,11 +1,13 @@
 const { createWorker } = require('tesseract.js');
 const { data } = require('./data.js')
 
-async function run(media_id, path) {
-    console.log(`running ocr on #${media_id} at ${path}`)
+async function run(mediaId, path) {
+    const mimeType = await data.getMimeType(mediaId)
+    if (mimeType == 'image/heic') return
+    console.log(`running ocr on #${mediaId} at ${path}`)
     const worker = await createWorker('eng')
     const ret = await worker.recognize(path)
-    await data.addOCRText(media_id, ret.data.text)
+    await data.addOCRText(mediaId, ret.data.text)
     await worker.terminate()
 }
